Make dict-strign codec and apply to HDF5
